nuked old frontend sort code minor refactoring
